Move shared types to shared package
