+++ packages/delta-server/src/delta/commands/PartitionCommands.ts
@@ -9,11 +9,12 @@ import {
     DeletePartitionCommand,
     DeltaEvent,
     PartitionAddedEvent,
-    PartitionDeletedEvent
+    PartitionDeletedEvent,
+    type ErrorDelta
 } from "@lionweb/server-delta-shared"
 import { deltaLogger } from "@lionweb/server-shared"
 import { DeltaContext } from "../DeltaContext.js"
-import { affectedNodeMessage, newErrorDelta, ErrorDelta, affectedPartitionMessage } from "../events.js"
+import { affectedNodeMessage, newErrorDelta, affectedPartitionMessage } from "../events.js"
 import { Participation } from "../participation/index.js"
 import { DeltaFunction } from "./DeltaUtil.js"
 import { validateProperTree } from "./Validations.js"
