+++ packages/delta-shared/src/DeltaUtil.ts
@@ -6,7 +6,7 @@ import {
     DeltaCommand,
     DeltaEvent,
     DeltaRequest,
-    DeltaResponse
+    DeltaResponse, ErrorEvent, ErrorResponse
 } from "./types/index.js"
 
 export type MessageFromClient = DeltaCommand | DeltaRequest | DeltaAdminRequest
@@ -14,6 +14,14 @@ export type MessageToClient = DeltaEvent | DeltaResponse | DeltaAdminResponse
 
 export function findDistributableInfos(delta: MessageFromClient | MessageToClient): AdditionalInfo[] {
     return delta.additionalInfos.filter(info => info.distribute)
-} 
+}
 
+export type ErrorDelta = ErrorEvent | ErrorResponse
+
+export function isErrorEvent(object: unknown): object is ErrorEvent {
+    return (object as ErrorEvent).messageKind === "ErrorEvent"
+}
+export function isErrorResponse(object: unknown): object is ErrorResponse {
+    return (object as ErrorResponse).messageKind === "ErrorResponse"
+}
